* moved generic init resume state tests out of pandas init tests into their own file
